stocky-ai/requirements.txt
python-telegram-bot==21.5
kiteconnect==5.0.1
aiosqlite==0.20.0
APScheduler==3.10.4
pyotp==2.9.0
requests==2.32.3
python-dotenv==1.2.1
yfinance==1.1.0
feedparser==6.0.12
